QA-914: Broke out into components and added highlighting to attribute
